# docs/openapi/openapi.yaml
openapi: 3.0.0
paths:
...
  /v1/auth/password/reset/request:
    post:
      description: Enqueues a password reset email for an existing user. Returns 204
        even if the email is unknown to avoid account enumeration.
      operationId: auth.password.reset.request
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequestDto"
      responses:
        "204":
          description: ""
      summary: Request password reset
      tags: *a1
      x-error-codes:
        - VALIDATION_FAILED
        - RATE_LIMITED
        - INTERNAL
...
  /v1/me/account-deletion/request:
    post:
      description: Schedules account deletion 30 days in the future. The account
        remains usable during the grace period and the request can be canceled.
        After the grace period, the account is de-identified (PII erased) and
        the email becomes reusable.
      operationId: users.me.accountDeletion.request
      parameters:
        - name: Idempotency-Key
          in: header
          description: "Idempotency key for safe retries of write requests. Replays return
            `Idempotency-Replayed: true`."
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 128
            description: "Opaque string (recommended: UUIDv4)."
      responses:
        "204":
          description: ""
      security:
        - access-token: []
      summary: Request account deletion (30-day grace)
      tags: &a6
        - Users
      x-error-codes:
        - UNAUTHORIZED
        - IDEMPOTENCY_IN_PROGRESS
        - CONFLICT
        - INTERNAL
...
  /v1/sync/daily-usage:
    post:
      description: "Uploads a batch of daily usage aggregates for a registered sync
        device. Requires Idempotency-Key. All-or-nothing validation; does not
        create devices. Limits: 1000 facts/request, 100 models/fact, 60
        pushes/15 min per user."
      operationId: sync.dailyUsage.push
      parameters:
        - name: Idempotency-Key
          in: header
          description: "Idempotency key for safe retries of write requests. Replays return
            `Idempotency-Replayed: true`."
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
            description: "Opaque string (recommended: UUIDv4)."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushDailyUsageRequestDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushDailyUsageEnvelopeDto"
      security:
        - access-token: []
      summary: Push daily usage facts
      tags:
        - Sync
      x-error-codes:
        - VALIDATION_FAILED
        - UNAUTHORIZED
        - FORBIDDEN
        - RATE_LIMITED
        - SYNC_CONTRACT_UNSUPPORTED
        - SYNC_DEVICE_NOT_FOUND
        - SYNC_IDENTITY_INVALID
        - SYNC_PAYLOAD_TOO_LARGE
        - IDEMPOTENCY_IN_PROGRESS
        - CONFLICT
        - INTERNAL
info:
  title: Burnly API
  description: Generated API contract (code-first).
  version: 0.1.0
  contact: {}
...
servers:
  - url: /
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: "First-party access token (Authorization: Bearer <token>)"
  schemas:
...
    PasswordResetRequestDto:
      type: object
      properties:
        email:
          type: string
          example: user@example.com
      required:
        - email
...
    SyncUsageWindowDto:
      type: object
      properties:
        startDate:
          type: string
          example: 2026-07-08
        endDate:
          type: string
          example: 2026-07-08
        scope:
          type: string
          enum:
            - rolling
          example: rolling
      required:
        - startDate
        - endDate
        - scope
    PushDailyUsageRequestDto:
      type: object
      properties:
        contractVersion:
          type: number
          example: 1
        clientDeviceId:
          type: string
          example: example-device-1
        appVersion:
          type: string
          example: 0.1.20
        reportingTimezone:
          type: string
          example: UTC
        clientRevision:
          type: number
          example: 1
          minimum: 1
        window:
          $ref: "#/components/schemas/SyncUsageWindowDto"
        facts:
          type: array
          items:
            type: object
            additionalProperties: true
          maxItems: 1000
          description: Daily usage facts; deep validation in service. Over-limit returns
            SYNC_PAYLOAD_TOO_LARGE.
      required:
        - contractVersion
        - clientDeviceId
        - appVersion
        - reportingTimezone
        - clientRevision
        - window
        - facts
    PushDailyUsageCountsDto:
      type: object
      properties:
        received:
          type: number
          example: 1
        upserted:
          type: number
          example: 1
        removed:
          type: number
          example: 0
        unchanged:
          type: number
          example: 0
        rejected:
          type: number
          example: 0
      required:
        - received
        - upserted
        - removed
        - unchanged
        - rejected
    PushDailyUsageResultDto:
      type: object
      properties:
        clientDeviceId:
          type: string
        acceptedAt:
          type: string
          example: 2026-07-09T12:00:00.000Z
        clientRevision:
          type: number
          example: 1
        window:
          $ref: "#/components/schemas/SyncUsageWindowDto"
        counts:
          $ref: "#/components/schemas/PushDailyUsageCountsDto"
      required:
        - clientDeviceId
        - acceptedAt
        - clientRevision
        - window
        - counts
    PushDailyUsageEnvelopeDto:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/PushDailyUsageResultDto"
      required:
        - data